Update Blip_Buffer.cpp for new.h error

+++ NesCore/nes_apu/Blip_Buffer.cpp
@@ -3,7 +3,7 @@
 #include "Blip_Buffer.h"
 
 #include <math.h>
-#include <new.h>
+#include <new>
 #include <string.h>
 
 /* Copyright (C) 2003-2005 Shay Green. This module is free software; you
@@ -44,7 +44,7 @@ void Blip_Buffer::clear(bool entire_buffer)
 
 blargg_err_t Blip_Buffer::sample_rate(long new_rate, int msec)
 {
-	unsigned new_size = (ULONG_MAX >> BLIP_BUFFER_ACCURACY) + 1 - widest_impulse_ - 64;
+	unsigned new_size = (UINT_MAX >> BLIP_BUFFER_ACCURACY) + 1 - widest_impulse_ - 64;
 	if (msec != blip_default_length)
 	{
 		size_t s = (new_rate * (msec + 1) + 999) / 1000;
